backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
requests==2.31.0
feedparser==6.0.10
google-generativeai==0.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.0
